fix: fix screen and window.

+++ bridge/core/frame/screen.d.ts
@@ -1,7 +1,7 @@
 import {EventTarget} from "../dom/events/event_target";
 
 export interface Screen extends EventTarget {
-  readonly availWidth: DartImpl<double>;
+  readonly availWidth: DartImpl<int64>;
   readonly availHeight: DartImpl<int64>;
   readonly width: DartImpl<int64>;
   readonly height: DartImpl<int64>;
